requirements.txt
absl-py==1.0.0
argcomplete==1.10.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.0
asteval==0.9.25
astunparse==1.6.3
attrs==21.2.0
backcall==0.2.0
beautifulsoup4==4.8.2
beep==2021.12.15.7
bleach==4.1.0
boto3==1.19.12
botocore==1.22.12
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
click==8.0.3
colorama==0.4.4
compressed-rtf==1.0.6
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
distlib==0.3.4
Django==4.0.2
docx2txt==0.8
ebcdic==1.1.1
EbookLib==0.17.1
entrypoints==0.3
extract-msg==0.28.7
filelock==3.6.0
Flask==2.0.3
Flask-Cors==3.0.10
flatbuffers==2.0
fonttools==4.28.5
future==0.18.2
gast==0.5.3
google-auth==2.6.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.44.0
h5py==3.5.0
idna==3.3
imageio==2.16.1
IMAPClient==2.1.0
importlib-metadata==4.11.2
ipykernel==6.6.0
ipython==7.30.1
ipython-genutils==0.2.0
itsdangerous==2.1.0
jedi==0.18.1
Jinja2==3.0.3
jmespath==0.10.0
joblib==1.1.0
jsonschema==4.3.1
jupyter-client==7.1.0
jupyter-core==4.9.1
jupyter-http-over-ws==0.0.8
jupyterlab-pygments==0.1.2
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==13.0.0
lmfit==1.0.3
lxml==4.7.1
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
monty==2021.12.1
msgpack-python==0.5.6
mysql-connector-python==8.0.27
nbclient==0.5.9
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.4
networkx==2.7.1
notebook==6.4.6
numpy==1.21.4
oauthlib==3.2.0
olefile==0.46
opencv-contrib-python==4.5.5.64
opencv-python==4.5.5.64
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.4
pandocfilters==1.5.0
parso==0.8.3
pdfminer.six==20191110
pickleshare==0.7.5
Pillow==8.4.0
platformdirs==2.5.1
playsound==1.3.0
prometheus-client==0.12.0
prompt-toolkit==3.0.24
protobuf==3.19.4
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.12.0
pydash==5.1.0
Pygments==2.10.0
pyparsing==3.0.6
pyrsistent==0.18.0
python-dateutil==2.8.2
python-docx==0.8.11
python-pptx==0.6.21
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.2.0
pywin32==302
pywinpty==1.1.6
pyzmq==22.3.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
s3transfer==0.5.0
scikit-image==0.19.2
scikit-learn==1.0.1
scipy==1.5.4
Send2Trash==1.8.0
six==1.12.0
sortedcontainers==2.4.0
soupsieve==2.3.1
SpeechRecognition==3.8.1
sqlparse==0.4.2
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
TextExtract==1.2.0
textract==1.6.4
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.0.0
tifffile==2022.2.9
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
typing_extensions==4.1.1
tzdata==2021.5
tzlocal==4.1
uncertainties==3.1.6
urllib3==1.26.7
virtualenv==20.13.2
virtualenvwrapper-win==1.2.7
watchtower==2.0.1
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.3
wrapt==1.13.3
xlrd==1.2.0
XlsxWriter==3.0.2
xmltodict==0.12.0
zipp==3.7.0
